Move quantization converters under config transforms
