+++ torchtitan/config/transform/__init__.py
@@ -10,6 +10,14 @@ from .apply import apply_transforms, transform_model_config_
 from .base import convert_config_type, ModelConfigTransform
 from .context_parallel import ContextParallelTransform
 from .lora import LoRAConverter
+from .quantization import (
+    Float8GroupedExpertsConverter,
+    Float8LinearConverter,
+    MXFP8GroupedExpertsConverter,
+    MXFP8LinearConverter,
+    NVFP4LinearConverter,
+    QuantizationConverter,
+)
 
 __all__ = [
     "ModelConfigTransform",
@@ -18,4 +26,10 @@ __all__ = [
     "convert_config_type",
     "ContextParallelTransform",
     "LoRAConverter",
+    "Float8GroupedExpertsConverter",
+    "Float8LinearConverter",
+    "MXFP8GroupedExpertsConverter",
+    "MXFP8LinearConverter",
+    "NVFP4LinearConverter",
+    "QuantizationConverter",
 ]
