+++ torchtitan/quantization/__init__.py
@@ -4,54 +4,15 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-# [Note] Getting the 'torchao' package:
-# This script requires the 'torchao' package to function correctly.
-# Please ensure you have this package installed from the appropriate repository.
-# You can obtain it from https://github.com/pytorch/ao by following the
-# installation instructions.
+"""Quantized module and tensor building blocks."""
 
-# Note: Performance
-# The quantization modules are intended to be ran under `torch.compile` for competitive performance
+from .float8 import Float8Linear
+from .mxfp8 import MXFP8Linear
+from .nvfp4 import NVFP4Linear
 
-from dataclasses import dataclass
-
-from torchtitan.protocols.model import ModelConfigConverter
-
-
-class QuantizationConverter(ModelConfigConverter):
-    """Base class for quantization converters.
-
-    Subclasses define a nested Config and implement ``convert()``
-    to transform the model config tree.
-    """
-
-    @dataclass(kw_only=True, slots=True)
-    class Config(ModelConfigConverter.Config):
-        model_compile_enabled: bool = False
-        """Whether torch.compile is enabled for the model."""
-
-
-# Re-export all public symbols so callers can import from the package directly.
-from .float8 import (  # noqa: F401, E402
-    Float8GroupedExpertsConverter,
-    Float8Linear,
-    Float8LinearConverter,
-)
-from .mxfp8 import (  # noqa: F401, E402
-    MXFP8GroupedExpertsConverter,
-    MXFP8Linear,
-    MXFP8LinearConverter,
-)
-from .nvfp4 import NVFP4Linear, NVFP4LinearConverter  # noqa: F401, E402
 
 __all__ = [
-    "Float8GroupedExpertsConverter",
     "Float8Linear",
-    "Float8LinearConverter",
-    "MXFP8GroupedExpertsConverter",
     "MXFP8Linear",
-    "MXFP8LinearConverter",
     "NVFP4Linear",
-    "NVFP4LinearConverter",
-    "QuantizationConverter",
 ]
